requirements.txt
fastapi==0.104.1
strawberry-graphql[fastapi]==0.216.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
msal==1.25.0
pydantic[email]
urllib3==2.0.7
requests==2.31.0